fast: Rename #hasField to #isEnabled
As pointed out by Ville Aine, Presence Map as defined in section 10.5
may indicate the presence of other entities than fields (e.g. Template
Identifier).

Moreover, the presence of a field may or may not be indicated by the
presence map depending on the field operator and optionality of field.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Reported-by: Ville Aine <ville.aine@iki.fi>